refactor: UserService 기능 병합
