+++ src/main/java/com/wow/delivery/service/UserService.java
@@ -2,6 +2,7 @@ package com.wow.delivery.service;
 
 import com.wow.delivery.entity.User;
 import com.wow.delivery.error.ErrorCode;
+import com.wow.delivery.error.exception.DuplicateException;
 import com.wow.delivery.error.exception.MismatchException;
 import com.wow.delivery.repository.UserRepository;
 import com.wow.delivery.util.PasswordEncoder;
@@ -13,20 +14,34 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final UserDuplicateCheckService userDuplicateCheckService;
 
     public void signup(User user) {
-        userDuplicateCheckService.isDuplicateUser(user);
+        validDuplicateUser(user);
         userRepository.save(user);
     }
 
     public void signin(String email, String password) {
-        User findUser = userRepository.findUserByEmail(email)
-                .orElseThrow(()-> new MismatchException(ErrorCode.MISMATCH_ACCOUNT));
-
+        User findUser = userRepository.getUserByEmail(email);
 
         if (!PasswordEncoder.matchesPassword(password, findUser.getPassword(), findUser.getSalt())) {
             throw new MismatchException(ErrorCode.MISMATCH_ACCOUNT);
         }
+    }
+
+    private void validDuplicateUser(User user) {
+        if (isDuplicateEmail(user.getEmail())) {
+            throw new DuplicateException(ErrorCode.DUPLICATE_EMAIL);
+        }
+        if (isDuplicatePhoneNumber(user.getPhoneNumber())) {
+            throw new DuplicateException(ErrorCode.DUPLICATE_PHONE_NUMBER);
+        }
+    }
+
+    private boolean isDuplicateEmail(String email) {
+        return userRepository.existsByEmail(email);
+    }
+
+    private boolean isDuplicatePhoneNumber(String phoneNumber) {
+        return userRepository.existsByPhoneNumber(phoneNumber);
     }
 }
